use the correct program to determine if a program exists; type -P
We are a bash script, so type -P does precisely what people think they
want `which` to do. Except `which` doesn't exist on my system! And in
general it is non-portable -- it has many different implementations,
some of which don't return failure on failures, and others of which
don't actually, erm, check $PATH (looking at you, ~/.cshrc `which`).